refactor: reorganize library routes and update test cases
