+++ routes/web.php
@@ -12,9 +12,12 @@ use App\Biblioteca\Controllers\AutoriController;
 use App\Biblioteca\Controllers\ClassificazioniController;
 use App\Biblioteca\Controllers\EditoriController;
 use App\Biblioteca\Controllers\EtichetteController;
+use App\Biblioteca\Controllers\LibriBorrowController;
 use App\Biblioteca\Controllers\LibriCollocazioneController;
 use App\Biblioteca\Controllers\LibriController;
+use App\Biblioteca\Controllers\LibriDeletedController;
 use App\Biblioteca\Controllers\LibriPrestitiController;
+use App\Biblioteca\Controllers\SearchableLibriController;
 use App\Biblioteca\Controllers\VideoController;
 use App\Nomadelfia\Azienda\Controllers\AziendeController;
 use App\Nomadelfia\Azienda\Controllers\AziendeLavoratoreController;
@@ -234,34 +237,35 @@ Route::prefix('scuola')->middleware('auth')->name('scuola.')->group(function () 
 
 Route::prefix('biblioteca')->middleware('auth')->group(function () {
     Route::view('/', 'biblioteca.index')->withoutMiddleware('auth')->name('biblioteca');
-    Route::get('libri', [LibriController::class, 'showSearchLibriForm'])->withoutMiddleware('auth')->name('libri.ricerca');
-    Route::get('libri/ricerca', [LibriController::class, 'searchConfirm'])->withoutMiddleware('auth')->name('libri.ricerca.submit');
-    Route::get('etichette/preview', [EtichetteController::class, 'preview'])->withoutMiddleware('auth')->name('libri.etichette.preview');
+    Route::get('books', [SearchableLibriController::class, 'index'])->withoutMiddleware('auth')->name('book.search.index');
+    Route::get('libri/search', [SearchableLibriController::class, 'search'])->withoutMiddleware('auth')->name('libri.ricerca.submit');
 
+    Route::get('libri/{idLibro}', [LibriController::class, 'show'])->middleware('can:biblioteca.libro.prenota')->name('libro.dettaglio');
     Route::get('libri/{idLibro}/collocazione', [LibriCollocazioneController::class, 'show'])->middleware('can:biblioteca.libro.visualizza')->name('libro.collocazione');
     Route::post('libri/{idLibro}/collocazione/update', [LibriCollocazioneController::class, 'updateCollocazione'])->middleware('can:biblioteca.libro.modifica')->name('libro.collocazione.update');
     Route::post('libri/{idLibro}/confirm', [LibriCollocazioneController::class, 'confirmCollocazione'])->middleware('can:biblioteca.libro.modifica')->name('libro.collocazione.update.confirm');
 
-    Route::get('libri/inserimento', [LibriController::class, 'showInsertLibroForm'])->middleware('can:biblioteca.libro.inserisci')->name('libri.inserisci');
-    Route::post('libri/inserimento', [LibriController::class, 'insertConfirm'])->middleware('can:biblioteca.libro.inserisci')->name('libri.inserisci.Confirm');
+    Route::get('libri/inserimento', [LibriController::class, 'create'])->middleware('can:biblioteca.libro.inserisci')->name('libri.inserisci');
+    Route::post('libri/inserimento', [LibriController::class, 'store'])->middleware('can:biblioteca.libro.inserisci')->name('libri.inserisci.Confirm');
+    Route::get('libri/{idLibro}/modifica', [LibriController::class, 'edit'])->middleware('can:biblioteca.libro.modifica')->name('libro.modifica');
+    Route::post('libri/{idLibro}/modifica', [LibriController::class, 'update'])->middleware('can:biblioteca.libro.modifica');
 
     Route::get('libri/prestiti', [LibriPrestitiController::class, 'view'])->middleware('can:biblioteca.libro.prenota')->name('libri.prestiti');
     Route::get('libri/prestiti/ricerca', [LibriPrestitiController::class, 'search'])->middleware('can:biblioteca.libro.visualizza')->name('libri.prestiti.ricerca');
     Route::get('libri/prestiti/{idPrestito}', [LibriPrestitiController::class, 'show'])->middleware('can:biblioteca.libro.visualizza')->name('libri.prestito');
-    Route::get('libri/prestiti/{idPrestito}/modifica', [LibriPrestitiController::class, 'edit'])->middleware('can:biblioteca.libro.modifica')->name('libri.prestito.modifica'); //->middleware('can:edit,App\Libro')->
-    Route::post('libri/prestiti/{idPrestito}/modifica', [LibriPrestitiController::class, 'editConfirm'])->middleware('can:biblioteca.libro.modifica'); //->name('libri.prestito.modificaConfirm');
+    Route::get('libri/prestiti/{idPrestito}/modifica', [LibriPrestitiController::class, 'edit'])->middleware('can:biblioteca.libro.modifica')->name('libri.prestito.modifica');
+    Route::post('libri/prestiti/{idPrestito}/modifica', [LibriPrestitiController::class, 'update'])->middleware('can:biblioteca.libro.modifica');
     Route::post('libri/prestiti/{idPrestito}/concludi', [LibriPrestitiController::class, 'conclude'])->middleware('can:biblioteca.libro.prenota')->name('libri.prestito.concludi');
 
-    Route::get('libri/eliminati', [LibriController::class, 'showDeleted'])->name('libri.eliminati');
-    Route::get('libri/{idLibro}', [LibriController::class, 'show'])->middleware('can:biblioteca.libro.prenota')->name('libro.dettaglio');
-    Route::get('libri/{idLibro}/modifica', [LibriController::class, 'edit'])->middleware('can:biblioteca.libro.modifica')->name('libro.modifica');
-    Route::post('libri/{idLibro}/modifica', [LibriController::class, 'editConfirm'])->middleware('can:biblioteca.libro.modifica');
-    Route::get('libri/{idLibro}/elimina', [LibriController::class, 'delete'])->middleware('can:biblioteca.libro.elimina')->name('libro.elimina');
-    Route::post('libri/{idLibro}/elimina', [LibriController::class, 'deleteConfirm'])->middleware('can:biblioteca.libro.elimina');
-    Route::get('libri/{idLibro}/prenota', [LibriController::class, 'book'])->middleware('can:biblioteca.libro.prenota')->name('libri.prenota');
-    Route::post('libri/{idLibro}/prenota', [LibriController::class, 'bookConfirm'])->middleware('can:biblioteca.libro.prenota');
-    Route::post('libri/{idLibro}/ripristina', [LibriController::class, 'restore'])->middleware('can:biblioteca.libro.elimina')->name('libri.ripristina');
+    Route::get('libri/{idLibro}/prenota', [LibriBorrowController::class, 'create'])->middleware('can:biblioteca.libro.prenota')->name('libri.prenota');
+    Route::post('libri/{idLibro}/prenota', [LibriBorrowController::class, 'store'])->middleware('can:biblioteca.libro.prenota');
 
+    Route::get('libri/eliminati', [LibriDeletedController::class, 'index'])->name('libri.eliminati');
+    Route::get('libri/{idLibro}/elimina', [LibriDeletedController::class, 'create'])->middleware('can:biblioteca.libro.elimina')->name('libro.elimina');
+    Route::delete('libri/{idLibro}/elimina', [LibriDeletedController::class, 'deleteConfirm'])->middleware('can:biblioteca.libro.elimina');
+    Route::put('libri/{idLibro}/ripristina', [LibriDeletedController::class, 'restore'])->middleware('can:biblioteca.libro.elimina')->name('libri.ripristina');
+
+    Route::get('etichette/preview', [EtichetteController::class, 'preview'])->withoutMiddleware('auth')->name('libri.etichette.preview');
     Route::get('etichette', [EtichetteController::class, 'view'])->middleware('can:biblioteca.etichetta.visualizza')->name('libri.etichette');
     Route::post('etichette', [EtichetteController::class, 'etichetteFromToCollocazione'])->middleware('can:biblioteca.etichetta.visualizza')->name('libri.etichette.aggiungi');
     Route::post('etichette/add/{idLibro}', [EtichetteController::class, 'addLibro'])->middleware('can:biblioteca.etichetta.inserisci')->name('libri.etichette.aggiungi.libro');
