--- tests/Biblioteca/Feature/HomePageTest.php
@@ -1,11 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Tests\Biblioteca\Feature;
-
-it('can render the homepage', function (): void {
-    $this
-        ->get('/biblioteca')
-        ->assertSee('Libri');
-});
